tests/requirements.txt
# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
